% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefactR-package.R
\docType{package}
\name{RefactR}
\alias{RefactR}
\alias{RefactR-package}
\title{RefactR - Refactor R code and tests}
\description{
Provides tools to automatically refactor R code, adding or extracting
functions, arguments, tests, documentation, etc.
}

